Adjusting ui code structure
代码结构大调整，UI创建部分移至ui文件夹中
重构和简化聚类、WD14部分的代码
工具类脚本统一移动至ui/tools
