+++ ui/tools/webui_error_wrapper.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Callable, Optional
+from typing import Callable
 
 """ (inspect.signature(func).parameters) 也可以获取输入数量 """
