--- venv_python.ps1
@@ -1,5 +1,0 @@
-.\venv\Scripts\activate
-
-python
-
-pause
